Update to Qt 6.4
First step towards running LogLite as a native process on ARM64 macOS, which is only natively supported in open-source Qt as of version 6.2.
Required changes were mostly renamed and obsolete flags. However, there was a bigger change to replace the deprecated QRegExp usage with QRegularExpression. In particular, the file path detection needs to be sufficiently tested now.