Fix Windows builds to work with Qt 6
Unfortunately this comes at the cost of the taskbar icon functionality. The relevant module was removed in Qt 6. However, there is a new module pending which is meant to provide this functionality in a cross-platform way.
Also, this removes the Qt 4 compatibility instructions that came in with the auto-generated CMakeLists.txt. There's room for more improvements on that front.